requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.36.0
Django==5.0.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==10.4.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
